wasi workaround: fix inexplicable issue w/ require order
This started happening around the time I rebased of 0.34 /
master sometime in late April 2020. Eventually had to play whack-a-mole here to add some requires needed to get the
spec suite generating. Something weird going on here with
ordering dependencies and compiling for WASI. src/file.cr is
the first one with issues, when it tries to sub-class File from IO::FileDescriptor (undefined constant)