Implement segfault handler in Crystal
Required adding some C bindings for various signal structures,
based on rust's fairly comprehensive libc bindings.

This removes the need to build and link a seperate C library and
reduces the complexity of cross-compiling.

siginfo_t can be fairly complex, involving large context-dependent
unions that vary across archiecture. In Crystal's case, we only care
about the segfault version of the structure, so we omit the others.